requirements.txt
connexion
flask
uwsgi
bcrypt
pexpect
pyjwt
cryptography
requests
